Style and data fetching improvements
After using this plugin for quite a while I've decided to add some extra features to it and make it more appealing so my wife and I can easily read it.

I've changed the email style and format, for every budget that is respected you will get a "green card" and for the ones that are not you will get a "red card". The red card will fetch all the transactions from that budget, group them by category and list them in descending order in that card so you know exactly what was spent on what. At the very end, you get a monthly report card that will give insight into how much you earned how much you spent and most important how much you saved and the percentage of your savings per month (in case you want to hit a specific percentage).

Please note that the "£" was hard-coded in my version but removed in this PR to allow other currencies to be used. There is some improvement that can be done on that side as well since you can fetch the currency from FireFly III.

As for future improvements I might add in the future an annual review card that you can enable in the config file to be added to the monthly report so you also have that functionality there.